Parte B/requirements.txt
certifi
charset-normalizer
click
colorama
contractions
emoji
et_xmlfile
filelock
fsspec
ftfy
huggingface-hub
idna
imbalanced-learn
iniconfig
Jinja2
joblib
lightgbm
MarkupSafe
mpmath
networkx
nltk
numpy
openpyxl
packaging
pandas
pluggy
pyahocorasick
pytest
python-dateutil
pytz
PyYAML
regex
requests
safetensors
scikit-learn
scipy
setuptools
six
sklearn-compat
sympy
textsearch
threadpoolctl
tokenizers
torch
tqdm
transformers
typing_extensions
tzdata
urllib3
wcwidth
xgboost